chat_translator/requirements.txt
openai-whisper
ffmpeg-python
transformers==4.41.2
librosa
pyttsx3
SpeechRecognition
PyAudio